# xts: avoid redundant bounds checks