Small changes to improve performance of xts by about 6%.

This removes all bounds checks, which can be verified by running:

  go build -gcflags="-d=ssa/check_bce/debug=1" .

Before this would show mutiple Found IsInBounds/IsSliceInBounds in the inner loops of Encrypt and Decrypt.

Tweaked the benchmark to use a larger buffer to make the improvement more visible. XTS is often used with disk sector sizes, so larger values are appropriate.

  goos: linux
  goarch: amd64
  pkg: golang.org/x/crypto/xts
  cpu: Intel(R) Xeon(R) W-2135 CPU @ 3.70GHz
         │ before.txt  │             after.txt              │
         │   sec/op    │   sec/op     vs base               │
  XTS-12   1.720µ ± 2%   1.619µ ± 2%  -5.87% (p=0.000 n=10)

         │  before.txt  │              after.txt              │
         │     B/s      │     B/s       vs base               │
  XTS-12   177.4Mi ± 2%   188.5Mi ± 2%  +6.24% (p=0.000 n=10)

         │ before.txt │           after.txt            │
         │    B/op    │    B/op     vs base            │
  XTS-12   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
  ¹ all samples are equal

         │ before.txt │           after.txt            │
         │ allocs/op  │ allocs/op   vs base            │
  XTS-12   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
  ¹ all samples are equal