@ncruces 2024-10-14
There's a larger gain to be had by unrolling the `range tweak` inner-loops (about 15% over this, 20% over the original), but that's more code, so maybe not worth it.